Use reader-server timestamp instead of firstSeen and lastSeen
The readers cannot sync their clocks with a time server, and programmatically setting the clock is not possible with LLRP (or not documented at least). Reader-server time will be good enough as long as the network lag between the reader and the server is low, and we don't do bath processing.